requirements.txt
libigl >= 2.4
numpy >= 1.24
usd-core >= 23.11
opencv-python >= 4.8.1.78
pillow >= 11.0.0
plyfile >= 1.1